Add FileWatcher init
