+++ electron/src/file_watcher/index.js
@@ -1,150 +1,55 @@
-import path from 'node:path';
-import { watch } from 'node:fs/promises';
+import { settingsEmitter } from "../settings.cjs";
+import { FileWatcher } from "./file_watcher.js";
 
-class TakeEvent extends Event {
-    constructor(take) {
-        super('takeUpdate');
-        this.take = take;
+let audioFilesPath;
+let trackName;
+let watcher;
+
+let notifyNewTake;
+
+export async function fileWatcherInit(mainWindow) {
+    settingsEmitter.on('change', handleNewSettings);
+    settingsEmitter.on('loaded', handleNewSettings);
+
+    notifyNewTake = (take) => {
+        mainWindow.webContent.send('set-take', take);
     }
+
+    return mainWindow;
 }
 
-/*
-https://stackoverflow.com/questions/3561493/is-there-a-regexp-escape-function-in-javascript
-*/
-function escapeRegex(string) {
-    return string.replace(/[/\-\\^$*+?.()|[\]{}]/g, '\\$&');
+function handleNewSettings(settings) {
+    if (settings.counterMode != 'fileWatcherMode') {
+        if (watcher) {
+            watcher.stopWatching();
+        }
+    }
+
+    handleFileWatcherMode(settings.fileWatcherMode);
 }
 
-export class FileWatcher extends EventTarget {
-    constructor(audioFilePath) {
-        super();
-
-        this.audioFilePath = audioFilePath;
-        this.isWatching = false;
+async function handleFileWatcherMode(fileWatcherSettings) {
+    if (!watcher) {
+        audioFilesPath = fileWatcherSettings.audioFilesPath; 
+        watcher = new FileWatcher(audioFilesPath);
+        watcher.addEventListener('takeUpdate', handleTakeChange);
     }
 
-    get currentTake() {
-        if (!this.lastAudioFile) {
-            return 0;
-        }
+    if (audioFilesPath != fileWatcherSettings.audioFilesPath) {
+        audioFilesPath = fileWatcherSettings.audioFilesPath; 
+        await watcher.changeAudioFilesPath(audioFilesPath);
+    } 
 
-        const lastTake = this.#parseTake(this.lastAudioFile);
-        if (lastTake == false) {
-            return 0;
-        }
+    if (!trackName || trackName != fileWatcherSettings.trackName) {
+        trackName = fileWatcherSettings.trackName;
+        await watcher.watchTrackName(trackName);
+    }
+} 
 
-        return lastTake;
+function handleTakeChange(event) {
+    if (!notifyNewTake) {
+        return;
     }
 
-    #parseTake(audioFileName) {
-        if (!this.trackNameRE) {
-            return false;
-        }
-        
-        if (!this.trackNameRE.test(audioFileName)) {
-            return false;
-        }
-
-        const match = audioFileName.match(this.trackNameRE);
-        if (!match[3]) {
-            return 1;
-        }
-
-        return Number(match[3]) + 1;
-    }
-
-    watchTrackName(trackName) {
-        return this.stopWatching()
-            .then(() => {
-                this.trackName = trackName;
-                const trackNameEscaped = escapeRegex(trackName);
-                this.trackNameRE = new RegExp(
-                    `(${trackNameEscaped})(\.?([0-9]+))?(_[0-9]+\.wav)`
-                );
-
-                this.watchAbort = new AbortController();
-                this.watchGenerator = watch(
-                    this.audioFilePath, 
-                    {  
-                        signal: this.watchAbort.signal 
-                    }
-                );
-
-                this.isWatching = true;
-                this.#nextWatchResult();
-            });
-    }
-
-    #nextWatchResult() {
-        this.currentWatchPromise = this.watchGenerator
-            .next()
-            .then((watchIteratorResult) => {
-                this.#handleWatchIteratorResult(watchIteratorResult);
-            })
-            .catch((err) => {
-                if (err.name == 'AbortError') {
-                    this.isWatching = false;
-                    return;
-                }
-                throw err;
-            });
-    }
-
-    #handleWatchIteratorResult(watchIteratorResult) {
-        if (watchIteratorResult.done) {
-            return;
-        }
-
-        const changeEvent = watchIteratorResult.value;
-        this.#handleAudioFilesChange(changeEvent);
-        this.#nextWatchResult();
-    }
-
-    #handleAudioFilesChange(changeEvent) {
-        if (changeEvent.eventType != 'rename') {
-            return;
-        }
-
-        const filename = changeEvent.filename;
-        const take = this.#parseTake(filename);
-        if (!take) {
-            return;
-        }
-
-        this.lastAudioFile = filename;
-        this.dispatchEvent(new TakeEvent(take));
-    }
-
-    stopWatching() {
-        if (this.isWatching) {
-            this.watchAbort.abort();
-        }
-        return this.nextUpdate();
-    }
-
-    nextUpdate() {
-        if (!this.isWatching) {
-            return Promise.resolve();
-        }
-
-        return new Promise((resolve, reject) => {
-            this.currentWatchPromise
-                .then(() => {
-                    resolve();
-                });
-        });
-    }
-
-    changeAudioFilesDirectory(directory) {
-        this.audioFilePath = directory;
-
-        if (!this.isWatching) {
-            return Promise.resolve();
-        }
-
-        return this.stopWatching()
-            .then(() => {
-                return this.watchTrackName(this.trackName);
-            })
-    }
+    notifyNewTake(event.take);    
 }
