+++ electron/src/main.cjs
@@ -6,6 +6,7 @@ const {
   alwaysOnTopInit 
 } = require('./settings.cjs');
 const { registerKeyboardShortcuts } = require('./shortcuts.cjs');
+const { fileWatcherInit } = require('./file_watcher/index.js');
 
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
 if (require('electron-squirrel-startup')) {
@@ -44,7 +45,8 @@ app
   .then(registerSettingsHandlers)
   .then(createWindow)
   .then(alwaysOnTopInit)
-  .then(registerKeyboardShortcuts);
+  .then(registerKeyboardShortcuts)
+  .then(fileWatcherInit);
 
 // Quit when all windows are closed, except on macOS. There, it's common
 // for applications and their menu bar to stay active until the user quits
